fx quant: make sure observer is inserted before a quantized output
Summary:

Before: if an output was marked as quantized, it could actually not
be quantized, if the previous node was not quantized.

After: if an output was marked as quantized, it will be quantized
regardless of the quantization status of the previous node.

Test Plan:

```
python test/test_quantization.py TestQuantizeFxOps.test_quant_output_always_observed
```

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 5c02d605edf3a7687120ec6383bd3fd893c05f24
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49420