Prevent accidentally writing old style ops
Pull Request resolved: https://github.com/pytorch/pytorch/pull/49510

Adding old style operators with out arguments will break XLA. This prevents that. See for background: https://fb.workplace.com/groups/pytorch.dev/permalink/809934446251704/

This is a temporary change that will prevent this breakage for the next couple of days until the problem is resolved for good.
It will be deleted in https://github.com/pytorch/pytorch/pull/49164 then.
ghstack-source-id: 118756437

Differential Revision: [D25599112](https://our.internmc.facebook.com/intern/diff/D25599112/)